+++ libgstd/gstd_event_handler.c
@@ -98,7 +98,7 @@ gstd_event_handler_set_property (GObject * object,
     {
       GstdICreator *creator;
 
-      self->receiver = g_value_get_object (value);
+      self->receiver = g_value_dup_object (value);
       GST_INFO_OBJECT (self, "Changed receiver to %p", self->receiver);
 
       creator = GSTD_ICREATOR (g_object_new (GSTD_TYPE_EVENT_CREATOR,